+++ src/contents/ui/main.qml
@@ -36,9 +36,15 @@ Item {
 
 	// ------------------------------------------------------------------------------------------------------------------------
 
-	Plasmoid.backgroundHints: plasmoid.configuration. containerLayoutTransparentBackgroundEnabled ? "NoBackground" : "StandardBackground"
 	Plasmoid.preferredRepresentation: Plasmoid.compactRepresentation
-	Plasmoid.compactRepresentation: ExchangeContainer { }
+	Plasmoid.compactRepresentation: ExchangeContainer {}
+
+	// If ConfigurableBackground is set, the we most likely run on Plasma 5.19+ and if so, we prefer using
+	// widget's background control features instead.
+	Plasmoid.backgroundHints: (typeof PlasmaCore.Types.ConfigurableBackground !== "undefined"
+		? PlasmaCore.Types.DefaultBackground | PlasmaCore.Types.ConfigurableBackground
+		: plasmoid.configuration.containerLayoutTransparentBackgroundEnabled ? PlasmaCore.Types.NoBackground : PlasmaCore.Types.DefaultBackground
+	)
 
 	// ------------------------------------------------------------------------------------------------------------------------